Rearranged JointControlInterface location
