+++ RR_Godot/Godot/scenes/main.tscn
@@ -121,7 +121,6 @@ anchor_bottom = 1.0
 custom_styles/panel = SubResource( 1 )
 
 [node name="TitleBar" type="Control" parent="UI"]
-editor/display_folded = true
 anchor_right = 1.0
 margin_bottom = 30.0
 rect_min_size = Vector2( 0, 30 )
